fix: fixed templates and tests
