+++ Example/Mocky/AutoMockable.swift
@@ -1,6 +1,9 @@
 //
+//  AutoMockable.swift
+//  Mocky_Example
+//
 //  Created by Andrzej Michnia on 23.10.2017.
-//  Copyright © 2017 SwiftyMocky. All rights reserved.
+//  Copyright © 2017 CocoaPods. All rights reserved.
 //
 
 import Foundation
