+++ Example/Podfile
@@ -1,8 +1,8 @@
+platform :ios, '8.3'
 use_frameworks!
 
 target 'Mocky_Example' do
     pod 'RxSwift'
-    pod 'Mocky/AutoMockable', :path => '../'
   target 'Mocky_Tests' do
     inherit! :search_paths
   	pod 'Mocky', :path => '../'
